backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
mysqlclient>=2.2.0
joblib>=1.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
